+++ src/pref.tcl
@@ -387,9 +387,11 @@ proc pref::Tab {tab {nt ""} {doit no}} {
       UpdateSyntaxTab
       UpdateSyntaxTab 2
     }
-    lassign [$obDl2 csGet $cs] fg - bg
-    [$obDl2 TexSample] configure -fg $fg -bg $bg
-    [$obDl2 TexCSample] configure -fg $fg -bg $bg
+    lassign [$obDl2 csGet $cs] fg - bg - - sbg sfg ibg
+    [$obDl2 TexSample] configure -fg $fg -bg $bg \
+      -selectbackground $sbg -selectforeground $sfg -insertbackground $ibg
+    [$obDl2 TexCSample] configure -fg $fg -bg $bg \
+      -selectbackground $sbg -selectforeground $sfg -insertbackground $ibg
     set data(INI,CSsaved) $cs
   }
   if {[string match root* $geo]} {
@@ -880,7 +882,7 @@ proc pref::InitSyntaxTcl {colornames} {
   set var "(Multiline string)
     Args=$args"
   winfo interps -displayof [lindex $args 0]
-  return $var ;#! TODO
+  return $var ;#! text of TODO
 }}}
   set wk [$obDl2 TexTclKeys]
   ::apave::bindToEvent $wk <FocusOut> alited::pref::UpdateSyntaxTab
@@ -1265,7 +1267,7 @@ proc pref::Tkcon_Default {} {
   set al(tkcon,cols) 100
   set al(tkcon,fsize) 13
   set al(tkcon,geo) +1+31
-  set al(tkcon,topmost) 0
+  set al(tkcon,topmost) [expr {!$al(IsWindows)}]
 }
 #_______________________